DEPL-11980 Investigate "OSError: [Errno 9] Bad file descriptor" when running XL Deploy in Docker
Jython is having issue while uploading file with the alpine linux base image. The fix is to use other distribution as base images. We tested it with jre-slim, it uses Debian as base image. The caveat is, it increases the size of the image by 100 MB and as the base image of this docker image is changed the impact will also be on the xl-se-docker-demo repository as it uses this image as base image. We need to fix those as well.